improve menu web code style
