+++ components/menu/index.web.tsx
@@ -5,38 +5,7 @@ import objectAssign from 'object-assign';
 import SubMenu from './SubMenu.web';
 import List from '../list';
 import Flex from '../flex';
-
-export interface DataItem {
-  label?: any;
-  value?: any;
-  children?: any;
-  isLeaf?: boolean;
-  [key: string]: any;
-}
-
-export interface SubDataItem {
-  label?: any;
-  value?: any;
-  [key: string]: any;
-}
-
-export interface MenuProps {
-  /** web only */
-  prefixCls?: string;
-  /** web only */
-  className?: string;
-  style?: React.CSSProperties;
-  data?: Array<DataItem>;
-  value?: Array<string>;
-  onChange?: Function;
-  level?: number;
-  height?: number;
-}
-
-export interface MenuState {
-  SubMenuData: Array<SubDataItem>;
-  firstValue?: any;
-}
+import { MenuProps, MenuState } from './propTypes';
 
 export default class Menu extends React.Component<MenuProps, MenuState> {
   static propTypes = {
@@ -53,41 +22,37 @@ export default class Menu extends React.Component<MenuProps, MenuState> {
 
   constructor(props) {
     super(props);
+    const { data, value, level } = props;
 
-    const { data, value, level } = this.props;
+    if (level !== 2) {
+      this.state = {
+        SubMenuData: data,
+      };
+      return;
+    }
+
     if (value.length > 0) {
-      if (level === 2) {
-        let SubMenuData = (data.filter((el) => {
-            return el.value === (value.length > 0 && value[0] || null);
-          }))[0].children || [];
-        this.state = {
-          SubMenuData,
-          firstValue: value[0] || '',
-        };
-      } else {
-        this.state = {
-          SubMenuData: data,
-        };
-      }
+      const SubMenuData = data.filter(
+        el => el.value === (value.length > 0 && value[0] || null)
+      )[0].children || [];
+      this.state = {
+        SubMenuData,
+        firstValue: value[0] || '',
+      };
+      return;
+    }
+
+    const SubMenuData = data[0].children || [];
+    if (data[0].isLeaf) {
+      this.state = {
+        SubMenuData: [],
+        firstValue: '',
+      };
     } else {
-      if (level === 2) {
-        let SubMenuData = data[0].children || [];
-        if (data[0].isLeaf) {
-          this.state = {
-            SubMenuData: [],
-            firstValue: '',
-          };
-        } else {
-          this.state = {
-            SubMenuData,
-            firstValue: data[0].value,
-          };
-        }
-      } else {
-        this.state = {
-          SubMenuData: data,
-        };
-      }
+      this.state = {
+        SubMenuData,
+        firstValue: data[0].value,
+      };
     }
   }
 
@@ -99,72 +64,70 @@ export default class Menu extends React.Component<MenuProps, MenuState> {
       }, () => {
         this.props.onChange([el.value]);
       });
-    } else {
-      this.setState({
-        firstValue: el.value,
-        SubMenuData: el.children || [],
-      });
     }
+    this.setState({
+      firstValue: el.value,
+      SubMenuData: el.children || [],
+    });
   };
 
   onClickSubMenuItem = (el) => {
+    const { level, onChange } = this.props;
     setTimeout(() => {
-      if (this.props.level === 2) {
-        this.props.onChange([this.state.firstValue, el.value]);
-      } else {
-        this.props.onChange([el.value]);
-      }
+      onChange(level === 2 ? [this.state.firstValue, el.value] : [el.value]);
     }, 300);
   };
 
   render() {
-    let { className, data, prefixCls, height, value, level, style } = this.props;
+    const { className, data, prefixCls, value, level, style } = this.props;
+    let { height } = this.props;
+
     if (!height) {
       height = document.documentElement.clientHeight / 2;
     }
 
-    let heightStyle = {
+    const heightStyle = {
       height: `${ Math.round(height / 44) * 44 - 1 }px`,
       overflowY: 'scroll',
     };
 
-    style = objectAssign(style || {}, heightStyle);
-
-    let { SubMenuData, firstValue } = this.state;
+    const { SubMenuData, firstValue } = this.state;
 
     const wrapCls = classNames({
       [prefixCls]: true,
-      [className]: className,
+      [className]: !!className,
     });
 
-    let listContent = [];
-    data.forEach((el, index) => {
-      listContent.push(<List.Item
+    const listContent = data.map((el, index) => (
+      <List.Item
         className={el.value === firstValue ? `${prefixCls}-selected` : ''}
         onClick={this.onClickListItem.bind(this, el)}
         key={`listitem-1-${index}`}
-      >{el.label}</List.Item>);
-    });
+      >
+        {el.label}
+      </List.Item>
+    ));
 
     return (
       <div
         className={wrapCls}
-        style={style}
+        style={objectAssign({}, style, heightStyle)}
       >
         <Flex align="top">
           {level === 2 ? (
-          <Flex.Item style={heightStyle}>
-            <List>
-              <List.Body>
-                {listContent}
-              </List.Body>
-            </List>
-          </Flex.Item>) : null}
+            <Flex.Item style={heightStyle}>
+              <List>
+                <List.Body>
+                  {listContent}
+                </List.Body>
+              </List>
+            </Flex.Item>
+          ) : null}
           <Flex.Item style={heightStyle}>
             <SubMenu
-              value={SubMenuData.filter((el) => {
-                return el.value === (value.length && value[value.length - 1]);
-              })}
+              value={SubMenuData.filter(
+                el => el.value === (value.length && value[value.length - 1])
+              )}
               data={SubMenuData}
               onChange={this.onClickSubMenuItem}
             />
